requirements.txt
scipy
numpy
tqdm
pypinyin
jieba
jieba_fast
cython
python-Levenshtein
